refactor: TOP-18 ActivityViewBindingDelegate를 core_ui/delegate 패키지로 이동
